+++ tensorprime.py
@@ -10,7 +10,7 @@ import saveload
 import math
 from config import config
 from log_helper import init_logger
-
+import os
 jnp_precision = jnp.float64
 
 def is_known_mersenne_prime(p):
@@ -276,12 +276,48 @@ def result_is_nine(signal, power_bit_array, n):
 primes = [2, 3, 5, 7, 13, 17, 19, 31, 61, 89, 107, 127, 521, 607, 1279, 2203, 2281, 3217, 4253, 4423, 9689, 9941, 11213, 19937, 21701, 23209, 44497, 86243, 110503, 132049, 216091,
                        756839, 859433, 1257787, 1398269, 2976221, 3021377, 6972593, 13466917, 20996011, 24036583, 25964951, 30402457, 32582657, 37156667, 42643801, 43112609, 57885161, 74207281, 77232917, 82589933]
 
-exponent = primes[19]
-siglen= max(1,int(jnp.log2(exponent/2.5)))
-bit_array, power_bit_array, weight_array = initialize_constants(
-  exponent, siglen)
-t1 = time.time()
-s = prptest(exponent, siglen, bit_array, power_bit_array, weight_array)
-t2 = time.time()
-n = (1<<exponent) - 1
-print(result_is_nine(s, power_bit_array, n))
+def given_N_get_max_p(N, float_type=64):
+  if float_type == 32:
+    Bmant = 24
+  if float_type == 64:
+    Bmant= 53
+  AsympConst = 0.6
+  ln2inv = 1.0/jnp.log(2.0)
+  ln_N     = jnp.log(1.0*N)
+  lnln_N   = jnp.log(ln_N)
+  l2_N     = ln2inv*ln_N
+  lnl2_N   = jnp.log(l2_N)
+  l2l2_N   = ln2inv*lnl2_N
+  lnlnln_N = jnp.log(lnln_N)
+  l2lnln_N = ln2inv*lnlnln_N
+    
+  Wbits = 0.5*( Bmant - AsympConst - 0.5*(l2_N + l2l2_N) - 1.5*(l2lnln_N) )
+  maxExp2 = Wbits*N
+  return maxExp2
+
+exponent = primes[4]
+os.environ["XLA_FLAGS"] = '--xla_gpu_cuda_data_dir=/home/I/.guix-profile/bin/ --xla_force_host_platform_device_count=8'
+def multiple_of_n_smaller_than(n,smaller_than):
+  multiples = []
+  i = 1
+  multiple = i * n
+  while(multiple<smaller_than):
+    multiples.append(multiple)
+    i += 1
+    multiple = i* n
+  return multiples
+
+siglens =  multiple_of_n_smaller_than(len(jax.devices()), exponent)
+
+for siglen in siglens:
+  try:
+    print("Trying ", siglen)
+    bit_array, power_bit_array, weight_array = initialize_constants(
+      exponent, siglen)
+    s = prptest(exponent, siglen, bit_array, power_bit_array, weight_array)
+    n = (1<<exponent) - 1
+    print(result_is_nine(s, power_bit_array, n))
+    break
+  except Exception as e:
+    print(e)
+    continue
